refactor: 💡 user 메소드이름변경에 의한 dto 이름 변경
