+++ src/main/java/gwasuwonshot/tutice/user/dto/response/GetAccountByLessonResponse.java
@@ -1,4 +1,4 @@
-package gwasuwonshot.tutice.lesson.dto.response.getAccountByLesson;
+package gwasuwonshot.tutice.user.dto.response;
 
 import gwasuwonshot.tutice.user.entity.Account;
 import lombok.*;
@@ -7,14 +7,14 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Builder
-public class GetLessonAccountResponse {
+public class GetAccountByLessonResponse {
     private Long idx;
     private String name;
     private String bank;
     private String number;
 
-    public static GetLessonAccountResponse of(Account account) {
-        return GetLessonAccountResponse.builder()
+    public static GetAccountByLessonResponse of(Account account) {
+        return GetAccountByLessonResponse.builder()
                 .idx(account.getIdx())
                 .name(account.getName())
                 .bank(account.getBank())
